@@ -14,18 +14,27 @@
   },
   "publisher": "SAP",
   "main": "dist/form.umd.js",
+  "module": "./dist/form.es.js",
+  "exports": {
+    ".": {
+      "import": "./dist/form.es.js",
+      "require": "./dist/form.umd.js"
+    },
+    "./dist/form.css": {
+      "import": "./dist/form.css",
+      "require": "./dist/form.css"
+    }
+  },
   "scripts": {
-    "build": "vue-cli-service build --target lib --name form src/index.js",
-    "build-dev": "vue-cli-service build --mode development",
-    "build-for-sample": "vue-cli-service build --target lib --name form src/index.js --dest sample-app/form",
-    "lint": "vue-cli-service lint",
+    "build": "vite build",
+    "build-dev": "vite build --mode development",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore",
     "prepublishOnly": "npm run build",
-    "serve": "vue-cli-service serve",
-    "test": "vue-cli-service test:unit -w 1"
+    "test": "jest -w 1"
   },
   "husky": {
     "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+      "commit-msg": "commitlint -V -E HUSKY_GIT_PARAMS"
     }
   },
   "commitlint": {
@@ -45,53 +54,45 @@
   "eslintConfig": {
     "env": {
       "jest": true,
-      "node": true
+      "node": true,
+      "es2022": true,
+      "document": true
     },
     "parserOptions": {
-      "parser": "babel-eslint"
+      "sourceType": "module"
     },
     "extends": [
-      "plugin:vue/essential",
       "eslint:recommended"
     ],
     "rules": {},
     "root": true
   },
   "dependencies": {
+    "lodash": "^4.17.21",
     "strip-ansi": "^6.0.0",
-    "vue": "2.6.12",
-    "vuetify": "2.3.10"
+    "vue": "^3.3.4",
+    "vuetify": "^3.3.6"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.22.5",
     "@commitlint/cli": "17.6.1",
     "@commitlint/config-conventional": "17.6.1",
-    "@vue/cli-plugin-babel": "5.0.8",
-    "@vue/cli-plugin-eslint": "4.5.19",
-    "@vue/cli-plugin-unit-jest": "4.5.19",
-    "@vue/cli-service": "4.5.12",
-    "@vue/test-utils": "1.2.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^26.6.3",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/test-utils": "2.4.0",
+    "@vue/vue3-jest": "^29.2.4",
+    "babel-jest": "^29.5.0",
     "coveralls": "2.13.3",
     "cz-conventional-changelog": "3.3.0",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
+    "eslint": "^8.44.0",
+    "eslint-plugin-vue": "^9.15.1",
     "husky": "6.0.0",
     "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
     "jest-junit": "^16.0.0",
     "jest-serializer-vue": "^3.1.0",
-    "sass": "1.37.2",
-    "sass-loader": "10.0.2",
-    "vue-cli-plugin-vuetify": "2.4.1",
-    "vue-jest": "3.0.7",
-    "vue-template-compiler": "2.6.12",
-    "vuetify-loader": "1.7.3",
-    "webpack": "4.46.0"
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
+    "jest-transform-stub": "^2.0.0",
+    "sass": "^1.63.6",
+    "vite": "^4.3.9",
+    "yorkie": "^2.0.0"
   }
 }
