@@ -1,6 +1,9 @@
-import { mount } from '@vue/test-utils';
-import Vue from 'vue';
-import Form from '../src/Form.vue';
+import { mount, enableAutoUnmount } from "@vue/test-utils";
+import { nextTick } from "vue";
+import { createVuetify } from 'vuetify';
+import * as components from 'vuetify/lib/components/index.mjs';
+import FormVue from "../src/Form.vue";
+import QuestionCheckbox from '../src/packages/QuestionCheckbox.vue';
 import utils from './utils';
 
 const questionCheckbox = [
@@ -83,17 +86,35 @@ const questionCheckboxCheckedForceDefault = [
     __ForceDefault: true
   }
 ];
-
+enableAutoUnmount(afterEach); //Ensures wrapper component gets cleaned up after each test
 describe('Question of type checkbox', () => {
-  test('Checkbox', async () => {
-    const wrapper = mount(Form, {});
-    wrapper.setProps({ questions: questionCheckbox });
-    await Vue.nextTick();
 
-    const citizenship = wrapper.find('div[role="listitem"]');
+  let vuetify
+
+  beforeEach(() => {
+    document.body.setAttribute("data-app", "true");
+    vuetify = new createVuetify({
+      components
+    });
+  })
+
+  test('Checkbox', async () => {
+    const wrapper = mount(FormVue, {
+      global: {
+        plugins: [vuetify],
+        components: {
+          'QuestionCheckbox': QuestionCheckbox
+        }
+      },
+      attachTo: document.body
+    });
+    wrapper.setProps({ questions: questionCheckbox });
+    await nextTick();
+
+    const citizenship = wrapper.findComponent({name: 'v-list-item'});
     citizenship.trigger('click');
 
-    await Vue.nextTick();
+    await nextTick();
 
     expect(wrapper.emitted().answered).toBeTruthy();
     const emittedLength = wrapper.emitted().answered.length;
@@ -103,14 +124,23 @@ describe('Question of type checkbox', () => {
   });
 
   test('Checkbox with choices as function', async () => {
-    const wrapper = mount(Form, {});
+    const wrapper = mount(FormVue, {
+      global: {
+        plugins: [vuetify],
+        components: {
+          'QuestionCheckbox': QuestionCheckbox
+        }
+      },
+      attachTo: document.body
+    });
     wrapper.setProps({ questions: questionCheckboxChoicesAsFunction });
-    await Vue.nextTick();
+    await nextTick();
+    await nextTick();
 
-    const citizenship = wrapper.find('div[role="listitem"]');
+    const citizenship = wrapper.findComponent({name: 'v-list-item'});
     citizenship.trigger('click');
 
-    await Vue.nextTick();
+    await nextTick();
     expect(wrapper.emitted().answered).toBeTruthy();
     const emittedLength = wrapper.emitted().answered.length;
     const answered = wrapper.emitted().answered[emittedLength-1];
@@ -119,9 +149,17 @@ describe('Question of type checkbox', () => {
   });
 
   test('Checkbox with choices as function that return dynamic values', async () => {
-    const wrapper = mount(Form, {});
+    const wrapper = mount(FormVue, {
+      global: {
+        plugins: [vuetify],
+        components: {
+          'QuestionCheckbox': QuestionCheckbox
+        }
+      },
+      attachTo: document.body
+    });
     wrapper.setProps({ questions: questionCheckboxDynamicChoicesAsFunction });
-    await Vue.nextTick();
+    await nextTick();
 
     const name = wrapper.find('input');
     const country = "A country";
@@ -131,25 +169,41 @@ describe('Question of type checkbox', () => {
     // wait to account for debounce
     await utils.sleep(300);
 
-    await Vue.nextTick();
+    await nextTick();
     expect(wrapper.props("questions")[1]._choices[2].value).toBe(country);
   });
 
   test('Checkbox with checked choice', async () => {
-    const wrapper = mount(Form, {});
+    const wrapper = mount(FormVue, {
+      global: {
+        plugins: [vuetify],
+        components: {
+          'QuestionCheckbox': QuestionCheckbox
+        }
+      },
+      attachTo: document.body
+    });
     wrapper.setProps({ questions: questionCheckboxChecked });
-    await Vue.nextTick();
-    await Vue.nextTick();
+    await nextTick();
+    await nextTick();
 
     expect(wrapper.props("questions")[0].answer[1]).toBe("China");
     expect(wrapper.props("questions")[0].answer).toHaveLength(2);
   });
 
   test('Checkbox with checked choice, but force default', async () => {
-    const wrapper = mount(Form, {});
+    const wrapper = mount(FormVue, {
+      global: {
+        plugins: [vuetify],
+        components: {
+          'QuestionCheckbox': QuestionCheckbox
+        }
+      },
+      attachTo: document.body
+    });
     wrapper.setProps({ questions: questionCheckboxCheckedForceDefault });
-    await Vue.nextTick();
-    await Vue.nextTick();
+    await nextTick();
+    await nextTick();
 
     expect(wrapper.props("questions")[0].answer[0]).toBe("Germany");
     expect(wrapper.props("questions")[0].answer).toHaveLength(1);
