@@ -1,7 +1,10 @@
-import { mount } from '@vue/test-utils';
-import Vue from 'vue';
-import Form from '../src/Form.vue';
-import QuestionDateTime from './QuestionDateTime.vue';
+import { mount, enableAutoUnmount } from "@vue/test-utils";
+import { nextTick } from "vue";
+import { createVuetify } from 'vuetify';
+import * as components from 'vuetify/components'
+import * as labsComponents from 'vuetify/labs/components'
+import FormVue from "../src/Form.vue";
+import QuestionDateTimePlugin from '../sample-plugin/src';
 
 const questionDate = [
   {
@@ -11,24 +14,36 @@ const questionDate = [
   }
 ];
 
-describe('Question of custom type date', () => {
-  test('Custom Date', async () => {
-    Vue.component('QuestionDateTime', QuestionDateTime);
-    await Vue.nextTick();
-    const datePlugin = {
-      questionType: "date",
-      component: QuestionDateTime
-    };
+enableAutoUnmount(afterEach);
 
-    const wrapper = mount(Form, {
-      stubs: {
-        'QuestionDateTime': QuestionDateTime
+describe('Question of custom type date', () => {
+  let vuetify
+
+  beforeEach(() => {
+    document.body.setAttribute("data-app", "true");
+    vuetify = new createVuetify({
+      components: {
+        ...components,
+        ...labsComponents
       }
     });
-    wrapper.vm.registerPlugin(datePlugin);
-    await Vue.nextTick();
+  })
+  test('Custom Date', async () => {
+
+    const options = {};
+    const wrapper = mount(FormVue, {
+      global: {
+        plugins: [vuetify, [QuestionDateTimePlugin, options]],
+      },
+      attachTo: document.body
+    });
+    await nextTick();
+
+    wrapper.vm.registerPlugin(options.plugin);
+
+    await nextTick();
     wrapper.setProps({ questions: questionDate });
-    await Vue.nextTick();
+    await nextTick();
     
     expect(wrapper.find('p.question-label > span.question-message').element.innerHTML).toBe("birthday");
   });
