@@ -1,7 +1,19 @@
 module.exports = {
   verbose: true,
+  testEnvironment: 'jest-environment-jsdom',
+  testEnvironmentOptions: {
+    customExportConditions: ["node", "node-addons"],
+ },
+  coverageProvider: 'v8',
   testRegex: "(\/__tests__\/(.*)\.(test|spec))\.[jt]sx?$",   /* eslint-disable-line */
   collectCoverage: true,
+  moduleNameMapper: {
+    '^vue$': '<rootDir>/node_modules/vue'
+  },
+  modulePathIgnorePatterns: [
+    "<rootDir>/sample-app",
+    "<rootDir>/sample-app-vite"
+  ],
   collectCoverageFrom: [
     "src/**/*.{js,vue}",
     "!**/node_modules/**",
@@ -29,22 +41,19 @@ module.exports = {
     "json"
   ],
   transformIgnorePatterns: [
-    "<rootDir>/node_modules/(?!(@sap-devx|vuetify|material-design-icons-iconfont|@mdi/font)/)"
+    "<rootDir>/node_modules/(?!(@sap-devx|vuetify|material-design-icons-iconfont|@mdi/font)/)",
   ],
   modulePaths: [
     "<rootDir>/src",
     "<rootDir>/node_modules"
   ],
   transform: {
-    ".*\\.(vue)$": "vue-jest",
-    '^.+\\.vue$': 'vue-jest',
+    ".*\\.(vue)$": "@vue/vue3-jest",
+    '^.+\\.vue$': '@vue/vue3-jest',
     '.+\\.(css|styl|less|sass|scss|svg|png|jpg|ttf|woff|woff2)$': 'jest-transform-stub',
-    '^.+\\.tsx?$': 'ts-jest',
-    "^.+\\.js$": "<rootDir>/node_modules/babel-jest"
+    "^.+\\.js$": "<rootDir>/node_modules/babel-jest",
+    "^.+\\.mjs$": "<rootDir>/node_modules/babel-jest"
   },
-  setupFiles: [
-    "<rootDir>/__tests__/setup.js"
-  ],
   snapshotSerializers: [
     "<rootDir>/node_modules/jest-serializer-vue"
   ],
