@@ -5,9 +5,9 @@
         <input type="checkbox" name="" :id="getId('input')" @change="onAnswerChanged" v-model="checkVal">
         <span class="question-hint">{{ question.title }}</span>
         <span class="question-hint" v-if="question.hint">
-          <v-tooltip top max-width="350px">
-            <template v-slot:activator="{ on }">
-              <v-icon v-on="on">mdi-help-circle-outline</v-icon>
+          <v-tooltip location="top" max-width="350px">
+            <template v-slot:activator="{ props }">
+              <v-icon v-bind="props">mdi-help-circle-outline</v-icon>
             </template>
             <span>{{ question.hint }}</span>
           </v-tooltip>
