@@ -1,8 +1,8 @@
-import Vue from 'vue';
-import Vuetify from 'vuetify';
+import {createVuetify} from 'vuetify';
+import * as components from 'vuetify/components';
+import * as directives from 'vuetify/directives';
 
-Vue.use(Vuetify, {
-});
-
-export default new Vuetify({
-});
+export default createVuetify({
+    components,
+    directives
+  });
